requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
numpy==1.26.4
pillow==10.3.0
opencv-python==4.10.0.84
scikit-learn==1.5.0
scipy==1.13.1
torch==2.3.1
torchvision==0.18.1
pyyaml==6.0.1
